proxy_manager/requirements.txt
aiohttp==3.9.3
pydantic==2.6.1
PyYAML==6.0.1
aiosqlite==0.19.0
python-socks[asyncio]==2.4.4
colorama==0.4.6
fake-useragent==1.4.0
